refactor: extract signupForm
