+++ src/app/auth/components/SignUp/SignUp.tsx
@@ -1,11 +1,9 @@
-import { useState, useEffect, useMemo } from 'react';
-import { SubmitHandler, useForm, useWatch } from 'react-hook-form';
-import queryString from 'query-string';
-import { auth } from '@internxt/lib';
+import { useState, useMemo } from 'react';
+import { SubmitHandler } from 'react-hook-form';
 import { Link } from 'react-router-dom';
-import { Info, WarningCircle } from '@phosphor-icons/react';
 import { Helmet } from 'react-helmet-async';
 import localStorageService, { STORAGE_KEYS } from 'app/core/services/local-storage.service';
+import queryString from 'query-string';
 
 import { useAppDispatch } from 'app/store/hooks';
 import { userActions, userThunks } from 'app/store/slices/user';
@@ -15,20 +13,16 @@ import navigationService from 'app/core/services/navigation.service';
 import { productsThunks } from 'app/store/slices/products';
 import { AppView, IFormValues } from 'app/core/types';
 import { referralsThunks } from 'app/store/slices/referrals';
-import TextInput from '../TextInput/TextInput';
-import PasswordInput from '../PasswordInput/PasswordInput';
-import Button from '../Button/Button';
-import testPasswordStrength from '@internxt/lib/dist/src/auth/testPasswordStrength';
-import PasswordStrengthIndicator from 'app/shared/components/PasswordStrengthIndicator';
 import { useSignUp } from './useSignUp';
 import { UserSettings } from '@internxt/sdk/dist/shared/types/userSettings';
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import authService, { getNewToken } from 'app/auth/services/auth.service';
 import PreparingWorkspaceAnimation from '../PreparingWorkspaceAnimation/PreparingWorkspaceAnimation';
 import paymentService from 'app/payment/services/payment.service';
-import { MAX_PASSWORD_LENGTH } from '../../../shared/components/ValidPassword';
 import { decryptPrivateKey } from 'app/crypto/services/keys.service';
 import analyticsService from 'app/analytics/services/analytics.service';
+import FormComponent from './FormComponent';
+import Button from '../Button/Button';
 
 export interface SignUpProps {
   location: {
@@ -39,7 +33,6 @@ export interface SignUpProps {
 
 function SignUp(props: SignUpProps): JSX.Element {
   const { translate } = useTranslationContext();
-  const [isValidPassword, setIsValidPassword] = useState(false);
 
   const qs = queryString.parse(navigationService.history.location.search);
   const autoSubmit = useMemo(
@@ -51,87 +44,13 @@ function SignUp(props: SignUpProps): JSX.Element {
     qs.register === 'activate' ? 'activate' : 'appsumo',
     hasReferrer ? String(qs.ref) : undefined,
   );
-  const hasEmailParam = (qs.email && auth.isValidEmail(qs.email as string)) || false;
 
-  const getInitialEmailValue = () => {
-    if (hasEmailParam) {
-      return qs.email as string;
-    }
-
-    if (autoSubmit.enabled && autoSubmit.credentials?.email) {
-      return autoSubmit.credentials.email;
-    }
-  };
-  const {
-    register,
-    formState: { errors, isValid },
-    handleSubmit,
-    control,
-    getValues,
-    resetField,
-  } = useForm<IFormValues>({
-    mode: 'onChange',
-    defaultValues: {
-      email: getInitialEmailValue(),
-      password: autoSubmit.enabled && autoSubmit.credentials ? autoSubmit.credentials.password : '',
-    },
-  });
   const dispatch = useAppDispatch();
-  const password = useWatch({ control, name: 'password', defaultValue: '' });
   const [signupError, setSignupError] = useState<Error | string>();
   const [showError, setShowError] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const [passwordState, setPasswordState] = useState<{
-    tag: 'error' | 'warning' | 'success';
-    label: string;
-  } | null>(null);
-  const [showPasswordIndicator, setShowPasswordIndicator] = useState(false);
+
   const showPreparingWorkspaceAnimation = useMemo(() => autoSubmit.enabled && !showError, [autoSubmit, showError]);
-
-  const formInputError = Object.values(errors)[0];
-
-  let bottomInfoError: null | string = null;
-
-  if (formInputError?.message) {
-    bottomInfoError = formInputError.message;
-  } else if (showError && signupError) {
-    bottomInfoError = signupError.toString();
-  }
-
-  useEffect(() => {
-    if (autoSubmit.enabled && autoSubmit.credentials) {
-      onSubmit(getValues());
-    }
-  }, []);
-
-  useEffect(() => {
-    if (password.length > 0) onChangeHandler(password);
-  }, [password]);
-
-  function onChangeHandler(input: string) {
-    setIsValidPassword(false);
-    if (input.length > MAX_PASSWORD_LENGTH) {
-      setPasswordState({ tag: 'error', label: translate('modals.changePasswordModal.errors.longPassword') });
-      return;
-    }
-
-    const result = testPasswordStrength(input, (qs.email as string) === undefined ? '' : (qs.email as string));
-
-    setIsValidPassword(result.valid);
-    if (!result.valid) {
-      setPasswordState({
-        tag: 'error',
-        label:
-          result.reason === 'NOT_COMPLEX_ENOUGH'
-            ? 'Password is not complex enough'
-            : 'Password has to be at least 8 characters long',
-      });
-    } else if (result.strength === 'medium') {
-      setPasswordState({ tag: 'warning', label: 'Password is weak' });
-    } else {
-      setPasswordState({ tag: 'success', label: 'Password is strong' });
-    }
-  }
 
   const onSubmit: SubmitHandler<IFormValues> = async (formData, event) => {
     const redeemCodeObject = autoSubmit.credentials && autoSubmit.credentials.redeemCodeObject;
@@ -201,9 +120,6 @@ function SignUp(props: SignUpProps): JSX.Element {
       errorService.reportError(err);
       const castedError = errorService.castError(err);
       setSignupError(castedError.message);
-      if (castedError.message.includes('is already registered')) {
-        resetField('password');
-      }
     } finally {
       setShowError(true);
     }
@@ -214,6 +130,23 @@ function SignUp(props: SignUpProps): JSX.Element {
 
     return currentParams.toString() ? '/login?' + currentParams.toString() : '/login';
   };
+
+  const View = ({ view }) => {
+    const views = {
+      signUp: (
+        <FormComponent
+          onSubmit={onSubmit}
+          isLoading={isLoading}
+          signupError={signupError}
+          autoSubmit={autoSubmit}
+          showError={showError}
+        />
+      ),
+    };
+
+    return views[view];
+  };
+
   return (
     <>
       <Helmet>
@@ -230,80 +163,14 @@ function SignUp(props: SignUpProps): JSX.Element {
           <PreparingWorkspaceAnimation />
         ) : (
           <div className="flex flex-col items-start space-y-5">
-            <h1 className="text-3xl font-medium">{translate('auth.signup.title')}</h1>
-
-            <form className="flex w-full flex-col space-y-2" onSubmit={handleSubmit(onSubmit)}>
-              <TextInput
-                placeholder={translate('auth.email') as string}
-                label="email"
-                type="email"
-                disabled={hasEmailParam}
-                autoComplete="email"
-                register={register}
-                required={true}
-                minLength={{ value: 1, message: 'Email must not be empty' }}
-                error={errors.email}
-              />
-
-              <label className="space-y-0.5">
-                <PasswordInput
-                  className={passwordState ? passwordState.tag : ''}
-                  placeholder={translate('auth.password')}
-                  label="password"
-                  autoComplete="new-password"
-                  maxLength={MAX_PASSWORD_LENGTH}
-                  register={register}
-                  onFocus={() => setShowPasswordIndicator(true)}
-                  required={true}
-                  error={errors.password}
-                />
-                {showPasswordIndicator && passwordState && password && (
-                  <PasswordStrengthIndicator
-                    className="pt-1"
-                    strength={passwordState.tag}
-                    label={passwordState.label}
-                  />
-                )}
-                {bottomInfoError && (
-                  <div className="flex flex-row items-start pt-1">
-                    <div className="flex h-5 flex-row items-center">
-                      <WarningCircle weight="fill" className="mr-1 h-4 text-red" />
-                    </div>
-                    <span className="font-base w-56 text-sm text-red">{bottomInfoError}</span>
-                  </div>
-                )}
-              </label>
-
-              <div className="flex space-x-2.5 rounded-lg bg-primary/10 p-3 pr-4 dark:bg-primary/20">
-                <Info size={20} className="shrink-0 text-primary" />
-                <p className="text-xs">
-                  {translate('auth.signup.info.normalText')}{' '}
-                  <span className="font-semibold">{translate('auth.signup.info.boldText')}</span>{' '}
-                  <span className="font-semibold text-primary underline">
-                    <a
-                      href="https://help.internxt.com/en/articles/8450457-how-do-i-create-a-backup-key"
-                      target="_blank"
-                    >
-                      {translate('auth.signup.info.cta')}
-                    </a>
-                  </span>
-                </p>
-              </div>
-
-              <Button
-                disabled={isLoading || !isValidPassword}
-                text={translate('auth.signup.title')}
-                disabledText={
-                  isValid && isValidPassword
-                    ? `${translate('auth.signup.encrypting')}...`
-                    : translate('auth.signup.title')
-                }
-                loading={isLoading}
-                style="button-primary"
-                className="w-full"
-              />
-            </form>
-
+            {/* <View view="signUp" /> */}
+            <FormComponent
+              onSubmit={onSubmit}
+              isLoading={isLoading}
+              signupError={signupError}
+              autoSubmit={autoSubmit}
+              showError={showError}
+            />
             <span className="mt-2 w-full text-xs text-gray-50">
               {translate('auth.terms1')}{' '}
               <a href="https://internxt.com/legal" target="_blank" className="text-xs text-gray-50 hover:text-gray-60">
